+++ gimnasio/src/main/java/ar/edu/unlam/tallerweb1/controladores/ControladorAdministrador.java
@@ -147,59 +147,9 @@ public class ControladorAdministrador {
 		modelo.put("listaSucursal", servicioSucursal.listarSucursales());
 		return new ModelAndView("agregarSucursal",modelo);
 	}
-	@RequestMapping(path="/actividadesAdmin")
-	public ModelAndView seleccionarSucursal() {
-		ModelMap modelo = new ModelMap();
-		modelo.put("listaSucursales",servicioSucursal.listarSucursales());
-		return new ModelAndView("actividadesAdministrador",modelo);
-	}
-	@RequestMapping(path="/{idSucursal}/listaActividadesSucursal")
-	public ModelAndView abmActividades(@PathVariable ("idSucursal")Long idSucursal) {
-		ModelMap modelo = new ModelMap();
-		modelo.put("listaSucursalActividades", servicioActividad.listarActividadesEnSucursal(idSucursal));
-		return new ModelAndView("listaActividadesAbm",modelo);
-	}
-	@RequestMapping(path="/{idSucursalActividad}/modificacionActividad")
-	public ModelAndView modificacionActividad(@PathVariable Long idSucursalActividad) {
-		ModelMap modelo = new ModelMap();
-		SucursalActividad SucursalActividadVacia = new SucursalActividad();
-		modelo.put("sucursalActividad",servicioActividad.traerActividadSucursal(idSucursalActividad));
-		modelo.put("sucursalActividadVacia", SucursalActividadVacia);
-		return new ModelAndView("modificacionActividad",modelo);
-	}	
-	@RequestMapping(path="/{idSucursalActividad}/modificacionactividad", method= RequestMethod.POST)
-		public ModelAndView establecerModificacionActividad(@ModelAttribute ("sucursalActividadVacia") SucursalActividad sucursalActividadUpdate,@PathVariable Long idSucursalActividad,@RequestParam("nombreActividad") String nombre) {
-		SucursalActividad sucursalActividadBdd= servicioActividad.traerActividadSucursal(idSucursalActividad);
-		servicioActividad.modificarActividad(sucursalActividadUpdate, sucursalActividadBdd);
-		return new ModelAndView("redirect:/{idSucursalActividad}/modificacionActividad"); 
-	}
-	@RequestMapping(path="/{idSucursal}/bajaActividad")
-	public ModelAndView eliminarSucursalActividad(@PathVariable("idSucursal")Long idSucursal,@RequestParam("idSucursalActividad") Long idSucursalActividad){
-		ModelMap modelo = new ModelMap();
-		String exito = "la actividad fue eliminada con exito";
-		SucursalActividad sucursalActividad = servicioActividad.traerActividadSucursal(idSucursalActividad);
-		servicioActividad.eliminarSucursalActividad(sucursalActividad);
-		modelo.put("exito", exito);
-		return new ModelAndView("redirect:/{idSucursal}/listaActividadesSucursal");
-	}
-
 	@RequestMapping(path = "/sucursal/nuevaSucursal", method = RequestMethod.POST)
 	public ModelAndView agregarNuevaSucursal(@ModelAttribute ("sucursalVacia") Sucursal sucursalNueva) {
 		servicioSucursal.agregarSucursal(sucursalNueva);
 		return new ModelAndView("redirect:/sucursalesadmin");
 }
-	@RequestMapping(path="actividad/nuevaActividadEnSucursal")
-	public ModelAndView agregarNuevaSucursalActividad() {
-		ModelMap modelo = new ModelMap();
-		SucursalActividad sucursalActividadVacia = new SucursalActividad();
-		modelo.put("listaSucursales", servicioSucursal.listarSucursales());
-		modelo.put("sucursalActividadVacia", sucursalActividadVacia);
-		modelo.put("listaActividades", servicioActividad.listaActividades());
-		return new ModelAndView("formNuevaActividadEnSucursal",modelo);
-	}
-	@RequestMapping(path="actividad/nuevaActividadEnSucursalProc",method = RequestMethod.POST)
-	public ModelAndView agregarNuevaSucActividad(@ModelAttribute ("sucursalActividadVacia") SucursalActividad sucursalActividadVacia) {
-		servicioActividad.agregarSucursalActividad(sucursalActividadVacia);
-		return new ModelAndView("redirect:/actividadesAdmin");
-	}
 }
